requirements.txt
pydub>=0.25.1
openai>=1.40.0
elevenlabs>=1.3.0
google-cloud-texttospeech>=2.16.3
python-dotenv>=1.0.1
